backend/requirements.txt
aerich==0.9.1
arq==0.26.3
asyncpg==0.30.0
bcrypt==4.0.1
defusedxml==0.7.1
fastapi==0.115.13
feedparser==6.0.11
gunicorn==23.0.0
httpx==0.28.1
passlib==1.7.4
psycopg2-binary==2.9.10
pydantic==2.11.7
PyJWT==2.9.0
python-jose==3.5.0
python-multipart==0.0.20
redis==5.3.0
tortoise-orm==0.25.1
uv==0.7.16
uvicorn==0.34.3
watchfiles==1.1.0
websockets==15.0.1